+++ cli/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased
 ----------
 - Added support for PGP encrypting emails
+- Switched to honoring system-wide instead of per-use configuration by
+  default
 - Made content type to use configurable via `--content-type` argument